cloud-file-service-rakai/backend/metadata-service/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==1.4.27
pydantic==2.0.3
python-dotenv==0.21.0  # Updated to meet pydantic-settings requirement
python-multipart==0.0.5
aiofiles==0.7.0
alembic==1.7.5
psycopg2-binary==2.9.6
pydantic-settings==2.0.1